Move NearestAtoms & MinDist into min_dist.py
